+++ feature/main/src/commonMain/kotlin/com/ondot/main/setting/navMap/NavMapSettingScreen.kt
@@ -31,6 +31,7 @@ import com.dh.ondot.presentation.ui.theme.SETTING_NAV_MAP
 import com.dh.ondot.presentation.ui.theme.SETTING_NAV_MAP_PROVIDER_GUIDE
 import com.dh.ondot.presentation.ui.theme.SETTING_NAV_MAP_PROVIDER_TITLE
 import com.dh.ondot.presentation.ui.theme.WORD_SAVE
+import com.ondot.designsystem.components.MapProviderList
 import com.ondot.designsystem.components.OnDotButton
 import com.ondot.designsystem.components.OnDotText
 import com.ondot.designsystem.components.TopBar
@@ -59,7 +60,6 @@ fun NavMapSettingScreen(
     viewModel: SettingViewModel = koinViewModel(),
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-    val interactionSource = remember { MutableInteractionSource() }
 
     LaunchedEffect(Unit) {
         viewModel.getUserMapProvider()
@@ -76,7 +76,6 @@ fun NavMapSettingScreen(
     NavMapSettingContent(
         mapProviders = uiState.mapProviders,
         selectedProvider = uiState.selectedProvider,
-        interactionSource = interactionSource,
         onBack = popScreen,
         onProviderClick = viewModel::updateSelectedProvider,
         onSaveClick = viewModel::updateMapProvider,
@@ -87,7 +86,6 @@ fun NavMapSettingScreen(
 fun NavMapSettingContent(
     mapProviders: List<MapProvider>,
     selectedProvider: MapProvider,
-    interactionSource: MutableInteractionSource,
     onBack: () -> Unit,
     onProviderClick: (MapProvider) -> Unit,
     onSaveClick: () -> Unit,
@@ -122,7 +120,6 @@ fun NavMapSettingContent(
             modifier = Modifier.weight(1f),
             mapProviders = mapProviders,
             selectedProvider = selectedProvider,
-            interactionSource = interactionSource,
             onProviderClick = onProviderClick,
         )
 
@@ -154,84 +151,5 @@ private fun TopBarSection(onBack: () -> Unit = {}) {
             color = Gray0,
             modifier = Modifier.padding(top = if (getPlatform() == ANDROID) 50.dp else 70.dp),
         )
-    }
-}
-
-@Composable
-private fun MapProviderList(
-    modifier: Modifier = Modifier,
-    mapProviders: List<MapProvider>,
-    selectedProvider: MapProvider,
-    interactionSource: MutableInteractionSource,
-    onProviderClick: (MapProvider) -> Unit,
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.Top,
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        mapProviders.forEachIndexed { index, provider ->
-            MapProviderItem(
-                provider = provider,
-                isSelected = selectedProvider == provider,
-                modifier =
-                    Modifier
-                        .weight(1f)
-                        .padding(bottom = 12.dp),
-                aspectRatio = 105f / 128f,
-                interactionSource = interactionSource,
-                onClick = { onProviderClick(provider) },
-            )
-        }
-    }
-}
-
-@Composable
-private fun MapProviderItem(
-    provider: MapProvider,
-    isSelected: Boolean,
-    modifier: Modifier = Modifier,
-    aspectRatio: Float,
-    interactionSource: MutableInteractionSource,
-    onClick: () -> Unit,
-) {
-    val imageResource =
-        when (provider) {
-            MapProvider.KAKAO -> painterResource(Res.drawable.ic_kakao_map)
-            MapProvider.NAVER -> painterResource(Res.drawable.ic_naver_map)
-            MapProvider.APPLE -> painterResource(Res.drawable.ic_apple_map)
-        }
-
-    Box(
-        modifier =
-            modifier
-                .background(Gray400, RoundedCornerShape(12.dp))
-                .clip(RoundedCornerShape(12.dp))
-                .border(1.dp, if (isSelected) Green600 else Gray400, RoundedCornerShape(12.dp))
-                .aspectRatio(aspectRatio)
-                .clickable(
-                    interactionSource = interactionSource,
-                    indication = null,
-                    onClick = onClick,
-                ).padding(horizontal = 24.dp, vertical = 20.dp),
-        contentAlignment = Alignment.Center,
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Image(
-                painter = imageResource,
-                contentDescription = null,
-                modifier = Modifier.aspectRatio(1f),
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            OnDotText(
-                text = provider.providerName,
-                style = OnDotTextStyle.BodyMediumR,
-                color = Gray0,
-            )
-        }
     }
 }
